Sting vs level teleport revisted
This reverses all of fb4b3b8deaaea7a34ddaef5091340ad4c5f4c49f
except for the fixes37.0 entry and does a better job in a cleaner
fashion.  If Sting is going to start glowing and "you materialize
on a different level" is pending, give the materialize message
before the glowing message.  Otherwise handle both stop-glowing
and/or you-materialize in the normal fashion.