Qt+X11 on OSX
Using hints/macosx10.10-qt,
| make QTDIR={path} WANT_WIN_X11=1 USE_XPM=1
now builds a combined tty+Qt+X11 binary.  Linking the X11 code
works a lot better when the Makefile tells the linker where the
X11 libraries are....

This reconciles some of the differences between macosx10.10 and
macosx10.10-qt but there are still lots more (including lack of
WANT_WIN_CURSES in the latter).  And now 10.10 has differences
with the half dozen or so other macosx10.* hints.