Qt Confusion icon
In the Qt status panel the six characteristics and the older
status conditions all have icons (similar to map tiles) drawn
above their values.  (The 3.6 era conditions that I added all
have blank icons and are in need of artwork.  3.7 conditions
aren't implemented.)

Int, Stun, and Conf all feature a brain oriented towards the
player's right shoulder.  The intelligence one is just a bare
brain, the stunned one features a black cloud over it, and the
confusion one has something over it that is cloud-like but
shaped differently from Stun as well has being white rather
than black; I'm not sure what that depicts.  This transposes
the Confusion icon so that it faces the player's left shoulder
instead of right.  Not a very emphatic suggestion of confusion
but seems a useful difference without requiring any artistic
competence.