create monster creating concealed mimic revisisted
commit 85c64860ea4a8016dfdf8f11bfdd587970328c91:
| [...] but fixing this specific case is trivial.

Not trivial enough to avoid getting the details wrong.  An old
commit log message (ce1f195f5e4b16e4c6c1a8a40250bc2337bb0b23,
June of 2006) claimed that this was fixed for bag of tricks but
that was for monsters in general; mimics could still be wrong.